sender: extract out exporter library to github.com/koding
