+++ go/src/github.com/koding/eventexporter/logexporter.go
@@ -1,0 +1,14 @@
+package eventexporter
+
+type LogExporter struct {
+	Events []*Event
+}
+
+func NewLogExporter() *LogExporter {
+	return &LogExporter{Events: []*Event{}}
+}
+
+func (l *LogExporter) Send(event *Event) error {
+	l.Events = append(l.Events, event)
+	return nil
+}
